day-1-intro-ingestion-migrate/event-hubs/producer/requirements.txt
azure-eventhub==5.10.1
pandas==1.3.2
numpy==1.21.2
python-dotenv==0.19.0
requests==2.26.0
faust==1.10.4
jsonschema==3.2.0
SQLAlchemy==1.4.23
